admin-web: merged model files
