+++ admin/src/components/DAGConfig.tsx
@@ -2,8 +2,8 @@ import { Box, Button, Stack } from '@mui/material';
 import React from 'react';
 import { GetDAGResponse } from '../api/DAG';
 import { DAGContext } from '../contexts/DAGContext';
-import { Config } from '../models/Config';
-import { Step } from '../models/Step';
+import { Config } from '../models';
+import { Step } from '../models';
 import ConfigEditor from './ConfigEditor';
 import ConfigInfoTable from './ConfigInfoTable';
 import ConfigPreview from './ConfigPreview';
