--- admin/src/models/Condition.tsx
@@ -1,4 +1,0 @@
-export type Condition = {
-  Condition: string;
-  Expected: string;
-};
