--- admin/src/models/DAG.tsx
@@ -1,29 +1,0 @@
-import { Condition } from './Condition';
-import { Step } from './Step';
-
-export type DAG = {
-  ConfigPath: string;
-  Name: string;
-  ScheduleExp: string[];
-  Group: string;
-  Tags: string[];
-  Description: string;
-  Env: string[];
-  LogDir: string;
-  HandlerOn: HandlerOn;
-  Steps: Step[];
-  HistRetentionDays: number;
-  Preconditions: Condition[];
-  MaxActiveRuns: number;
-  Params: string[];
-  DefaultParams: string;
-  Delay: number;
-  MaxCleanUpTime: number;
-};
-
-export type HandlerOn = {
-  Failure: Step;
-  Success: Step;
-  Cancel: Step;
-  Exit: Step;
-};
