--- admin/src/models/StatusFile.tsx
@@ -1,6 +1,0 @@
-import { Status } from './Status';
-
-export type StatusFile = {
-  File: string;
-  Status: Status;
-};
